refactor: extract selector normalisation and url rebasing helpers
