+++ packages/beasties/src/index.ts
@@ -24,16 +24,12 @@ import path from 'node:path'
 
 import { applyMarkedSelectors, markOnly, parseStylesheet, serializeStylesheet, validateMediaQuery, walkStyleRules, walkStyleRulesWithReverseMirror } from './css'
 import { createDocument, serializeDocument } from './dom'
+import { isAlwaysCriticalSelector, normalizeCssSelector } from './selectors'
+import { REMOTE_URL_RE, resolveCssUrl, rewriteCssUrls } from './urls'
 import { createLogger, isSubpath } from './util'
 
-const removePseudoClassesAndElementsPattern = /(?<!\\)::?[a-z-]+(?:\(.+\))?/gi
-const implicitUniversalPattern = /([>+~])\s*(?!\1)([>+~])/g
-const emptyCombinatorPattern = /([>+~])\s*(?=\1|$)/g
-const removeTrailingCommasPattern = /\(\s*,|,\s*\)/g
 const LEADING_SLASH_OR_QUERY_RE = /^\/(?!\/)|[?#].*$/g
 const PUBLIC_PATH_RE = /(^\/(?!\/)|\/$)/g
-const REMOTE_URL_RE = /^https?:\/\//
-const BEFORE_AFTER_PSEUDO_RE = /^::?(?:before|after)$/
 const FONT_FAMILY_RE = /\bfont(?:-family)?\b/i
 // eslint-disable-next-line regexp/no-useless-assertions
 const BEASTIES_COMMENT_RE = /^(?<!! )beasties:(.*)/
@@ -41,10 +37,6 @@ const LEADING_SLASH_RE = /^\//
 const WHITESPACE_RE = /\s+/
 // eslint-disable-next-line regexp/no-super-linear-backtracking,regexp/no-misleading-capturing-group
 const URL_RE = /url\s*\(\s*(['"]?)(.+?)\1\s*\)/
-// unquoted urls cannot contain unescaped parentheses, and excluding them keeps
-// backtracking linear on input like `url((((...`
-const URL_RE_G = /url\((?:'([^']*)'|"([^"]*)"|([^()]*))\)/gi
-const ABSOLUTE_URL_RE = /^(?:[a-z][\w+.-]*:|\/\/|\/|#)/i
 
 interface PreFetchedStylesheet {
   link: ChildNode
@@ -605,12 +597,7 @@ export default class Beasties {
 
             // Strip pseudo-elements and pseudo-classes, since we only care that their associated elements exist.
             // This means any selector for a pseudo-element or having a pseudo-class will be inlined if the rest of the selector matches.
-            if (
-              sel === ':root'
-              || sel === 'html'
-              || sel === 'body'
-              || (sel[0] === ':' && BEFORE_AFTER_PSEUDO_RE.test(sel))
-            ) {
+            if (isAlwaysCriticalSelector(sel)) {
               return true
             }
 
@@ -788,56 +775,12 @@ export default class Beasties {
       return normalizedSelector
     }
 
-    normalizedSelector = sel
-      .replace(removePseudoClassesAndElementsPattern, '')
-      .replace(removeTrailingCommasPattern, match => (match.includes('(') ? '(' : ')'))
-      .replace(implicitUniversalPattern, '$1 * $2')
-      .replace(emptyCombinatorPattern, '$1 *')
-      .trim()
+    normalizedSelector = normalizeCssSelector(sel)
 
     this.#selectorCache.set(sel, normalizedSelector)
 
     return normalizedSelector
   }
-}
-
-/**
- * Resolve a `url()` value that is relative to `baseHref` (the location of the
- * stylesheet it was declared in) so that it can be used from the document instead.
- */
-function resolveCssUrl(url: string, baseHref: string): string {
-  if (!url || ABSOLUTE_URL_RE.test(url)) {
-    return url
-  }
-
-  const base = baseHref.split('?')[0]!.split('#')[0]!
-
-  if (REMOTE_URL_RE.test(base) || base.startsWith('//')) {
-    try {
-      const resolved = new URL(url, base.startsWith('//') ? `https:${base}` : base)
-      return base.startsWith('//') ? resolved.href.replace(REMOTE_URL_RE, '//') : resolved.href
-    }
-    catch {
-      return url
-    }
-  }
-
-  const dir = path.posix.dirname(base)
-  // the stylesheet sits alongside the document, so relative urls already resolve correctly
-  if (dir === '.' || dir === '') {
-    return url
-  }
-
-  return path.posix.join(dir, url)
-}
-
-function rewriteCssUrls(css: string, baseHref: string): string {
-  return css.replace(URL_RE_G, (match, singleQuoted?: string, doubleQuoted?: string, bare?: string) => {
-    const quote = singleQuoted !== undefined ? '\'' : doubleQuoted !== undefined ? '"' : ''
-    const url = singleQuoted ?? doubleQuoted ?? bare?.trim() ?? ''
-    const resolved = resolveCssUrl(url, baseHref)
-    return resolved === url ? match : `url(${quote}${resolved}${quote})`
-  })
 }
 
 function formatSize(size: number) {
